'Nother major commit.
Adding a launcher. Next commit will hopefully be support for multiple
maps. Networking is being put off because I want to integrate it with
physics, which requires physics to be completed first, which requires
Nazek42 to get some work done because I don't want to have to deal with
PyBox2D